+++ src/server/api/twilio_webhook_test.rs
@@ -50,7 +50,19 @@ async fn test_twilio_webhook_post_handler_success() {
         .body(body)
         .unwrap();
 
-    let response = app.oneshot(request).await.unwrap();
+    let response = app.clone().oneshot(request).await.unwrap();
 
     assert_eq!(response.status(), StatusCode::OK);
+
+    // Test with media
+    let body_media = Body::from("From=whatsapp%3A%2B14155238886&To=whatsapp%3A%2B1234567890&Body=Check+out+this+cake&NumMedia=1&MediaUrl0=https%3A%2F%2Fexample.com%2Fimage.jpg&MediaContentType0=image%2Fjpeg");
+    let request_media = Request::builder()
+        .method("POST")
+        .uri("/api/v1/webhooks/twilio")
+        .header("Content-Type", "application/x-www-form-urlencoded")
+        .body(body_media)
+        .unwrap();
+
+    let response_media = app.oneshot(request_media).await.unwrap();
+    assert_eq!(response_media.status(), StatusCode::OK);
 }
