+++ src/server/utils/BUILD.bazel
@@ -17,6 +17,7 @@ rust_library(
         "tier_middleware.rs",
         "edge_caching_middleware.rs",
         "tenant_middleware.rs",
+        "url.rs",
     ],
     crate_root = "mod.rs",
     edition = "2024",
