+++ src/server/utils/url.rs
@@ -1,0 +1,43 @@
+pub fn url_decode(input: &str) -> String {
+    let mut decoded = String::new();
+    let mut chars = input.chars().peekable();
+    while let Some(c) = chars.next() {
+        if c == '+' {
+            decoded.push(' ');
+        } else if c == '%' {
+            let mut hex = String::new();
+            if let Some(h1) = chars.next() {
+                hex.push(h1);
+                if let Some(h2) = chars.next() {
+                    hex.push(h2);
+                    if let Ok(byte) = u8::from_str_radix(&hex, 16) {
+                        decoded.push(byte as char);
+                    } else {
+                        decoded.push('%');
+                        decoded.push_str(&hex);
+                    }
+                } else {
+                    decoded.push('%');
+                    decoded.push(h1);
+                }
+            } else {
+                decoded.push('%');
+            }
+        } else {
+            decoded.push(c);
+        }
+    }
+    decoded
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn test_url_decode() {
+        assert_eq!(url_decode("Hello+World"), "Hello World");
+        assert_eq!(url_decode("Hello%20World"), "Hello World");
+        assert_eq!(url_decode("whatsapp%3A%2B1234567890"), "whatsapp:+1234567890");
+    }
+}
